Remove ServerReload brought in also by #1482/WFCORE-1460, and unify the two
