+++ testsuite/standalone/src/test/java/org/wildfly/core/test/standalone/mgmt/ReloadWithConfigTestCase.java
@@ -37,7 +37,7 @@ import javax.inject.Inject;
 
 import org.jboss.as.controller.PathAddress;
 import org.jboss.as.controller.operations.common.Util;
-import org.jboss.as.test.shared.ServerReload;
+import org.jboss.as.test.integration.management.util.ServerReload;
 import org.jboss.dmr.ModelNode;
 import org.junit.Assert;
 import org.junit.Test;
